+++ app/routes/legal.tsx
@@ -1,0 +1,2 @@
+export { loader } from './legal/_layout';
+export { default } from './legal/_layout';